+++ src/components/app-bar/app-bar.tsx
@@ -1,9 +1,0 @@
-import { Component } from '@stencil/core'
-
-@Component({
-  tag: 'slate-app-bar',
-  styleUrl: './app-bar.css',
-})
-class AppBar {}
-
-export { AppBar }