+++ src/components/index.ts
@@ -1,2 +1,0 @@
-export * from './container/container'
-export * from './stack/stack'