+++ src/tag/tag.js
@@ -1,0 +1,9 @@
+import { SlateElement } from '../slate-element.js'
+
+class Tag extends SlateElement {}
+
+if (!customElements.get('slate-tag')) {
+  customElements.define('slate-tag', Tag)
+}
+
+export { Tag }